refactor(pools): move type definitions from base-contract file totypes files
