+++ src/pools/erc20-collateral-pool.ts
@@ -1,23 +1,20 @@
 import { ethers } from 'ethers'
 
 import { miscErc20CollateralPool } from '../artifacts'
-import {
-  AdminFunctions,
-  BaseContract,
-  Erc20CollateralTokenPoolDetail,
-  Pagination,
-  Views
-} from '../base-classes'
+import { BaseContract } from '../base-classes'
 import { ecpErrorMessage } from '../errors'
 import {
   Borrow,
+  Erc20CollateralTokenPoolDetail,
   Functions,
   Lend,
   Pool,
   PoolInput,
-  PoolLiquidationInfo
+  PoolLiquidationInfo,
+  ERC20CollateralPoolViews as Views
 } from '../types/erc20-collateral-token'
-import { Abi, PrivateKey } from '../types/types'
+import { AdminFunctions } from '../types/pools'
+import { Abi, Pagination, PrivateKey } from '../types/types'
 import { NULL_ADDRESS, Role } from '../utilities/util'
 
 export class ERC20CollateralPool
