+++ src/pools/pools.ts
@@ -1,24 +1,21 @@
 import { ethers } from 'ethers'
 
 import { miscPools } from '../artifacts'
+import { BaseContract } from '../base-classes'
+import { cppErrorMessage, poolCommonErrorMessage } from '../errors'
+import { Erc20CollateralTokenPoolDetail } from '../types/erc20-collateral-token'
 import {
   AdminFunctions,
-  BaseContract,
-  Erc20CollateralTokenPoolDetail,
-  Pagination,
-  Views
-} from '../base-classes'
-import { cppErrorMessage, poolCommonErrorMessage } from '../errors'
-import {
   ContractPool,
   Functions,
   Pool,
   PoolCommit,
   PoolInput,
   PoolStatus,
-  PoolStatusOption
+  PoolStatusOption,
+  PoolViews as Views
 } from '../types/pools'
-import { Abi, PrivateKey } from '../types/types'
+import { Abi, Pagination, PrivateKey } from '../types/types'
 import { Role, getUnixEpochTime } from '../utilities/util'
 
 export class Pools
